% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anchor_delta.R
\name{compute_anchor_delta}
\alias{compute_anchor_delta}
\title{Compute the Anchor Change Score}
\usage{
compute_anchor_delta(
  dat = NULL,
  subject.id = NULL,
  time.var = NULL,
  anchor = NULL
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{subject.id}{variable of the subject ID in the dataframe numeric; default is USUBJID}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{anchor}{numeric anchor variable, default is PGIS}
}
\value{
returns a dataframe with two additional variables, anchor score at baseline and anchor change score
}
\description{
Compute the Anchor Change Score
}
