% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anchor_group.R
\name{compute_anchor_group}
\alias{compute_anchor_group}
\title{Compute the Anchor Group}
\usage{
compute_anchor_group(
  dat,
  anchor.variable = "PGIS_delta",
  number.of.anchor.groups = 5
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.variable}{numeric variable name that you want to use to create the anchor groups variable of the subject ID in the dataframe numeric; default is USUBJID}

\item{number.of.anchor.groups}{numeric, 3 or 5. If it's 5, all subjects with a non-missing value in the anchor.variable
will be assigned to: c('Improved_2+', 'Improved_1', 'Maintained', 'Deteriorated_1', 'Deteriorated_2+'). If number if 3,
then only c( 'Improved_1', 'Maintained', 'Deteriorated_1') will be used, with others set to NA.}
}
\value{
returns a dataframe with additional variable \code{anchor.group}
}
\description{
Compute the Anchor Group
}
