% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_change_score.R
\name{compute_change_score}
\alias{compute_change_score}
\title{Compute the PRO Change Score}
\usage{
compute_change_score(
  dat = NULL,
  subject.id = NULL,
  time.var = NULL,
  score = NULL
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{subject.id}{variable of the subject ID in the dataframe numeric; default is USUBJID}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{score}{numeric PRO score variable, default is Y_comp}
}
\value{
returns a dataframe with two additional variables, PRO score at baseline and PRO change score
}
\description{
Compute the PRO Change Score
}
