% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_icc21.R
\name{compute_icc21}
\alias{compute_icc21}
\title{Test-Retest Reliability}
\usage{
compute_icc21(
  dat,
  PRO.score = NULL,
  time.var = NULL,
  subject.id = NULL,
  anchor = NULL,
  stable.score = NULL,
  first.timepoint = NULL,
  second.timepoint = NULL
)
}
\arguments{
\item{dat}{dataframe with values}

\item{PRO.score}{specify the COA/PRO score you want to compute ICC of.
Can pass multiple scores, e.g. \code{PRO.score = c('Y_comp', 'Y_mcar', 'Y_mar')}}

\item{time.var}{the time variable, should be an ordered factor}

\item{subject.id}{the subject ID, should be a character variable}

\item{anchor}{this is the anchor used to condition on the stable subset of subjects}

\item{stable.score}{this is the anchor score that indicates the subject is stable.
For example, use \code{anchor = 'PGIS_delta'} and \code{stable.score = 0} to select
the subset of stable sujects.
Another example would be \code{anchor = 'anchor.groups'}
and \code{stable.score = 'Maintained'}.}

\item{first.timepoint}{specify the first timepoint to be used in the TRTR
computation; default is the first timepoint in your \code{time.var} variable}

\item{second.timepoint}{specifiy the second timepoint to be used in the TRTR
computation; default is the final timepoint in your \code{time.var} variable}
}
\value{
returns a table \code{icc.21} with the values
}
\description{
Uses the \code{ICC()} function in the \code{psych} package to compute the ICC(2,1)
on a stable subset of subjects.
}
