% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_known_groups_validity.R
\name{compute_known_groups_validity}
\alias{compute_known_groups_validity}
\title{Known-Groups Validity}
\usage{
compute_known_groups_validity(
  dat = NULL,
  PRO.score = NULL,
  time.var = NULL,
  val.var = NULL
)
}
\arguments{
\item{dat}{dataframe with values}

\item{PRO.score}{specify the COA/PRO score you want to compute ICC of.
Can only pass one character variable at a time.}

\item{time.var}{the time variable, should be an ordered factor}

\item{val.var}{specify the validator variables (reference measures).
Can only pass one variable at a time.}
}
\value{
returns two different footnotes to use in the table,
descriptive statistics table, and the model output
}
\description{
Known-Groups Validity
}
