% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prop_surp.R
\name{compute_prop_surp}
\alias{compute_prop_surp}
\title{Proportion of Subjects Surpassing the Threshold}
\usage{
compute_prop_surp(
  dat = NULL,
  anchor.group = NULL,
  time.var = NULL,
  timepoint = NULL,
  change.score = NULL,
  mean.or.median = NULL,
  threshold.label = NULL
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.group}{variable of the anchor group you want used here}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{timepoint}{default here is to just use the final timepoint, e.g., "Time_4". Please be sure to have
ordered your time variable correctly.}

\item{change.score}{the PRO change score you want the threshold computed on}

\item{mean.or.median}{pick a threshold type}

\item{threshold.label}{optional numeric value, threshold to pass;
default is default is "Improved_1" if "Improved" and "Deteriorated_1" if "Deteriorated"}

\item{threshold.type}{character, either Improved or Deteriorated}
}
\value{
list containing a (1) a table of the proportion of subjects in each anchor group
that surpass the threshold, and (2) a character vector to be optionally used as the caption in the eCDF plot
}
\description{
Computes the proportion of subjects surpassing the threshold,
either improvement or deterioration, in each anchor group.
This can then be passed as a caption to the eCDF plot.
}
