% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_thresholds.R
\name{compute_thresholds}
\alias{compute_thresholds}
\title{Compute Thresholds}
\usage{
compute_thresholds(
  dat = NULL,
  anchor.group = NULL,
  time.var = NULL,
  timepoint = NULL,
  change.score = NULL
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.group}{variable of the anchor group you want used here}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{timepoint}{default here is to just use the final timepoint, e.g., "Time_4". Please be sure to have
ordered your time variable correctly.}

\item{change.score}{the PRO change score you want the threshold computed on}
}
\value{
returns a dataframe with two additional variables, PRO score at baseline and PRO change score
}
\description{
Compute Thresholds
}
