% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_validity.R
\name{compute_validity}
\alias{compute_validity}
\title{Concurrent Validity}
\usage{
compute_validity(dat, PRO.score = NULL, time.var = NULL, val.var = NULL)
}
\arguments{
\item{dat}{dataframe with values}

\item{PRO.score}{specify the COA/PRO score you want to compute ICC of.
Can pass multiple scores, e.g. \code{PRO.score = c('Y_comp', 'Y_mcar', 'Y_mar')}}

\item{time.var}{the time variable, should be an ordered factor}

\item{val.var}{specify the validator variables (reference measures)
Can pass a character string, e.g., \code{val.var = c('PGIS', 'Val_1', 'Val_2')}}
}
\value{
returns a dataframe with the correlations
}
\description{
Uses the \code{polycor} package to compute the correlations
Pulls the baseline data, checks the variable type, selects
the correct correlation type
}
