% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_eCDF.R
\name{ggplot2_eCDF}
\alias{ggplot2_eCDF}
\title{Plot eCDF using \code{ggplot2}}
\usage{
ggplot2_eCDF(
  dat = NULL,
  output.plots = "both",
  anchor.group = NULL,
  time.var = NULL,
  change.score = NULL,
  print.to.png = FALSE,
  file.name = "Example_eCDF_ePDF_ggplot2",
  shell.table = FALSE
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{output.plots}{specify what you want printed out; 'eCDF', 'ePDF', or 'both'.
Default is 'both'}

\item{anchor.group}{variable of the anchor group you want used here}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{change.score}{indicate the name of the PRO change score}

\item{print.to.png}{logical, output the plots to a png file}

\item{file.name}{pass the name of the png file to print out}

\item{shell.table}{logical, if true the sample size will be "xx", can be used as shell tables}
}
\value{
Plot of eCDF using ggplot2 package
}
\description{
Plot the eCDF consistent with FDA COA Guidance 3, using \code{ggplot2} R package
}
\details{
Also requires \code{gridExtra}
}
