% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_eCDF_v2.R
\name{ggplot2_eCDF_v2}
\alias{ggplot2_eCDF_v2}
\title{Plot eCDF using \code{ggplot2}}
\usage{
ggplot2_eCDF_v2(
  dat = NULL,
  smoothed.ecdf = F,
  anchor.group = NULL,
  anchor.group.colors = NULL,
  time.var = NULL,
  change.score = NULL,
  shell.table = FALSE
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{smoothed.ecdf}{if TRUE, uses the smoothed density to plot the eCDF; default is FALSE}

\item{anchor.group}{variable of the anchor group you want used here}

\item{anchor.group.colors}{if you have unique anchor groups, you need to assign
colors to those anchor groups, such as c('Deteriorated_1' = 'red', 'Improved_1' = 'green')}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{change.score}{indicate the name of the PRO change score}

\item{shell.table}{logical, if true the sample size will be "xx", can be used as shell tables}
}
\value{
Plot of eCDF using ggplot2 package
}
\description{
Revised version 2
}
\details{
Plot the eCDF consistent with FDA COA Guidance 3, using \code{ggplot2} R package
}
