% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_ePDF.R
\name{ggplot2_ePDF}
\alias{ggplot2_ePDF}
\title{Plot ePDF (aka Histograms) using \code{ggplot2}}
\usage{
ggplot2_ePDF(
  dat = NULL,
  anchor.group = NULL,
  anchor.group.colors = NULL,
  time.var = NULL,
  change.score = NULL,
  shell.table = FALSE
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.group}{variable of the anchor group you want used here}

\item{anchor.group.colors}{if you have unique anchor groups, you need to assign
colors to those anchor groups, such as c('Deteriorated_1' = 'red', 'Improved_1' = 'green')}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{change.score}{indicate the name of the PRO change score}

\item{shell.table}{logical, if true the sample size will be "xx", can be used as shell tables}
}
\value{
Plot of ePDF using ggplot2 package
}
\description{
Plot the ePDF (histogram) consistent with FDA COA Guidance 3, using \code{ggplot2} R package
Generally presented along with the eCDF
}
