% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Plots with shared legend eCDF using \code{ggplot2}}
\usage{
grid_arrange_shared_legend(...)
}
\arguments{
\item{pass}{the plots pass the dataframe}
}
\value{
ggplot2 with a shared legend
}
\description{
Create multiple \code{ggplot2} plots with a shared legend
}
\details{
This is borrowed directly from https://rpubs.com/sjackman/grid_arrange_shared_legend
It requires \code{ggplot2}, \code{grid}, and \code{gridExtra}
}
