% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecdf.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot eCDF}
\usage{
plot_ecdf(
  dat = NULL,
  anchor.group = NULL,
  time.var = NULL,
  change.score = NULL,
  plot.title = "Empirical Cumulative Distribution of Change in Score from Baseline",
  plot.xlab = "Change in score from baseline",
  plot.ylab = "Cumulative proportion of patients",
  threshold.value = NULL,
  file.name = "Example_eCDF",
  shell.table = FALSE
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.group}{variable of the anchor group you want used here}

\item{time.var}{variable of the Time in the dataframe - PLEASE CHECK THAT THIS IS CORRECTLY ORDERED, default is Time}

\item{change.score}{indicate the name of the PRO change score}

\item{plot.title}{character vector to use as the plot title}

\item{plot.xlab}{character vector to pass to the xlab in base R plot functions}

\item{plot.ylab}{character vector to pass to the ylab in base R plot functions}

\item{file.name}{pass the name of the png file to print out}

\item{shell.table}{logical, if true the sample size will be "xx", can be used as shell tables}
}
\value{
Plot of eCDF
}
\description{
Plot the eCDF consistent with FDA COA Guidance 3, using base R plot functions.
This will be deprecated soon; inferior to \code{ggplot2} version.
}
