% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_irt_item.R
\name{sim_irt_item}
\alias{sim_irt_item}
\title{Simulate Items using IRT model}
\usage{
sim_irt_item(
  dat,
  J = NULL,
  K = NULL,
  item.param = NULL,
  latent.variable = "Y_comp",
  time.var = "Time"
)
}
\arguments{
\item{dat}{dataframe with values}

\item{J}{number of items}

\item{K}{number of categories}

\item{item.param}{option to pass the matrix of item parameters; you have to
name the slope "slope" and the intercepts "intercept_1", "intercept_2" etc
if you leave this empty, item parameters will be used}

\item{latent.variable}{pass the latent variable that will be used in the
IRT model to simulate the item responses}

\item{time.var}{specify the time variable, this is used to standardize the
latent variable based on the distribution at the first timepoint. That is,
the latent variable at timepoint 1 will have a mean of zero and sd = 1}

\item{anchor}{this is the anchor used to condition on the stable subset of subjects}
}
\value{
returns dataframe with J items in it, also returns the item parameters
and a separate dataframe with just the item responses
}
\description{
Pass the function a latent variable and IRT item parameters, the function
will simulate J items with K categories
}
