% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pro_dat.R
\name{sim_pro_dat}
\alias{sim_pro_dat}
\title{Simulate Longitudinal PRO Data with anchor and validator variables included}
\usage{
sim_pro_dat(
  N = 1000,
  number.timepoints = 4,
  number.anchor.categories = 4,
  reg.formula = formula(~Time + Group + Time * Group),
  Beta.anchor = 0,
  thresholds = c(0.2, 0.4, 0.6, 0.8),
  polychor.struc = "ar1",
  polychor.value = 0.4,
  number.of.anchor.groups = 5,
  Beta.PRO = NULL,
  corr = "ar1",
  cor.value = 0.8,
  var.values = 2,
  n.val = 5,
  n.cat = c(3, NA, NA, NA, 5),
  cor.val.ref = c(0.5, 0.6, 0.8, 0.2, 0.3)
)
}
\arguments{
\item{number.anchor.categories}{anchor generation - Pass to \code{sim_dat_ord()}: the number of PGIS anchor categories to be generated,
this is passed to the \code{number.groups} in the \code{sim_dat_ord} function}

\item{reg.formula}{anchor generation - Pass to \code{sim_dat_ord()}: this is a regression formula to pass to the anchor data generation; null is
formula(~ Group + Time + Time*Group),}

\item{Beta.anchor}{anchor generation - Pass to \code{sim_dat_ord()}: this is the Beta for the regression equation; numeric matrix of Beta values OR a scalar value != 0
that will be the final value of the interaction parameters, default is all(Beta == 0) for Type I error simulations}

\item{thresholds}{anchor generation - Pass to \code{sim_dat_ord()}: the ordinal generation uses a probit approach, so these are the thresholds}

\item{polychor.struc}{anchor generation - Pass to \code{sim_dat_ord()}: options for correlation structure  c('ind', 'ar1', 'cs'), default is 'ar1'}

\item{polychor.value}{anchor generation - Pass to \code{sim_dat_ord()}: numeric, the first corr in ar1, the corr in cs option}

\item{corr}{PRO data generation, options for correlation structure  c('ind', 'ar1', 'cs'), default is 'ar1'}

\item{cor.value}{PRO data generation, numeric, the first corr in ar1, the corr in cs option, default is 0.8}

\item{var.values}{PRO data generation, numeric, if numeric vector, the var values in
the var/covar matrix, if scalar, the final var value in the var/covar matrix, others filled in automatically}

\item{n.val}{Validator variable (Reference Measure) generation - Pass to \code{sim_val_var()}:
numeric, number of validators}

\item{n.cat}{Validator variable (Reference Measure) generation - Pass to \code{sim_val_var()}:
numeric vector,  number of categories in each validator,
NA for continuous validator variables}

\item{cor.val.ref}{Validator variable (Reference Measure) generation - Pass to \code{sim_val_var()}:
numeric vector, the correlation between the PRO score and each of the \code{n.val} validators}
}
\value{
returns a dataframe containing the simulated data
}
\description{
Simulate Longitudinal PRO Data with anchor and validator variables included
}
