% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_val_var.R
\name{sim_val_var}
\alias{sim_val_var}
\title{Simulate Longitudinal Ordinal Data}
\usage{
sim_val_var(
  dat,
  n.val = 5,
  n.cat = c(3, NA, NA, NA, 5),
  cor.val.ref = c(0.5, 0.6, 0.8, 0.2, 0.3)
)
}
\arguments{
\item{dat}{pass the dataframe}

\item{n.val}{numeric,  number of validators}

\item{n.cat}{numeric vector,  number of categories in each validator, NA for continuous validator variables}

\item{cor.val.ref}{Reference Measure generation, numeric vector, the correlation between the PRO score and each of the \code{n.val} validators}
}
\value{
returns a dataframe containing the simulated data
}
\description{
Simulate Longitudinal Ordinal Data
}
