% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anchor_group.R
\name{compute_anchor_group}
\alias{compute_anchor_group}
\title{Compute the Anchor Group}
\usage{
compute_anchor_group(dat, anchor.variable = "PGIS_delta")
}
\arguments{
\item{dat}{pass the dataframe}

\item{anchor.variable}{numeric variable name that you want to use to create the anchor groups variable of the subject ID in the dataframe numeric; default is USUBJID}
}
\value{
returns a dataframe with additional variable \code{anchor.group}
}
\description{
Compute the Anchor Group
}
